requirements.txt
# Libraries

grog
speechRecognition
pyttsx3
pyaudio
python-dotenv
streamlit
